exec: go files split
